+++ test/stubs/modules/relay_controller.h
@@ -33,6 +33,21 @@ static inline bool RelayController_isTimedEventActive() {
     return _stub_timed_event_active;
 }
 
+static inline void RelayController_requestRelease() {
+    if (!_stub_timed_event_active) _stub_timed_event_trigger_count++;
+    _stub_timed_event_active = true;
+}
+
+static inline bool RelayController_requestReleaseOff(bool surfaceSafe) {
+    if (!surfaceSafe) return false;
+    _stub_timed_event_active = false;
+    return true;
+}
+
+static inline bool RelayController_isReleaseActive() {
+    return _stub_timed_event_active;
+}
+
 static inline void RelayController_update() {}
 
 static inline void RelayController_emergencyDisable() {